screenshot_tool/requirements.txt
mss
Pillow
pyperclip
opencv-python
numpy
pyautogui
pytesseract
